// page-objects/base/time-out.enum.ts
// <copyright file='time-out' company='symplr'>
// Copyright © 2020 symplr. All rights reserved. Confidential and Proprietary.
// </copyright>

/**
 * Time out in ms
 *
 * export
 * enum {number}
 */
export enum TIME_OUT {
  xxs = 100,
  xs = 250,
  s = 500,
  m = 1000,
  l = 2500,
  xl = 5000,
  xxl = 10000,
  xxxl = 20000,
}